lab 6 work on next part
did not work something with the server not reading the env file correctly and had to change or restart this over to see if can get to work. not sure how to fix as this is a repo that was taken from my github account and just none of the credentials worked.